fix incorrect annotation Transient in java.Beans to javax.persistence

+++ src/main/java/com/unitoken/resume/database/DbTemplate.java
@@ -165,7 +165,7 @@ public class DbTemplate {
                     }
                 }, keyHolder);
                 if (rows == 1) {
-                    mapper.id.setter.invoke(bean, keyHolder.getKey());
+                    mapper.id.setter.invoke(bean, keyHolder.getKey().longValue());
                 }
             } else {
                 // id is specified
